Update [core] [File Lister] [malware scanner] [main window] [settings] [utils]
[core]
Add small malware db update system
Add python script for malware db update

[File Lister]
Roll back to old version

[malware scanner]
Fix tab select on python script error

[main window]
Notify for malware db update every 7 days (one week)

[settings]
Fix GIF bug: file already there

[utils]
Optimized fadeout and fadein effects

+++ ProHack Security Lite/SecLite_Settings.vb
@@ -304,7 +304,9 @@ Public Class SecLite_Settings
                 Dim imgpathArray As String() = Split(imgpath, "\")
                 Dim imgpathArrLen As Integer = imgpathArray.Length()
                 Dim imagename As String = imgpathArray(imgpathArrLen - 1)
-                System.IO.File.Copy(imgpath, wallpaperDir & imagename)
+                If Not System.IO.File.Exists(wallpaperDir & imagename) Then
+                    System.IO.File.Copy(imgpath, wallpaperDir & imagename)
+                End If
                 utils.core_Wallpaper = imagename
                 picPreview_Wallpaper.Image = Image.FromFile(wallpaperDir & imagename)
                 rbtn_Wallpaper_Image.ForeColor = blue
